+++ cases/testcases/unused_ret_vals/red1.clj
@@ -1,0 +1,9 @@
+(ns testcases.unused-ret-vals.red1
+  (:import
+   (java.io File)))
+
+(defn foo []
+  ;; Exercise `:unused-ret-vals` (static method call):
+  (Class/forName "a")
+
+  3)