+++ cases/testcases/unused_ret_vals/red3.clj
@@ -1,0 +1,9 @@
+(ns testcases.unused-ret-vals.red3
+  (:import
+   (java.io File)))
+
+(defn foo []
+  ;; Exercise `:unused-ret-vals` (instance method call):
+  (-> "afs" .getBytes String. .toString)
+
+  42)