+++ cases/testcases/unused_ret_vals/red4.clj
@@ -1,0 +1,10 @@
+(ns testcases.unused-ret-vals.red4
+  (:import
+   (java.io File)))
+
+(defn foo []
+  (try
+    ;; Exercise `:unused-ret-vals-in-try` (instance method call):
+    (-> "afs" .getBytes String. .toString)
+    1
+    (catch Exception _)))